Fix timestamp with positive system timezone offset
When no timestamp is provided the get_localized_timestamp function
attempts to calculate the current time in UTC and then adjust it
to the local timezone using the system's timezone offset.
When the local time is ahead of UTC, the previous implementation
was wrapping the time around as .seconds does not allow for a negative
difference.

Also swapped the deprecated `utcfromtimestamp(now)` with
`fromtimestamp(now, timezone.utc)`.