Send calculated UTC offset to clock
Previously the UTC offset was set to 0 and the UTC timestamp was
adjusted unless the user had set `tz_offset`.
Now, the UTC offset is calculated based on the local timezone.
For users in timezones with a non whole hour offset, we fallback to
sending the partial hour offset as adjusted UTC.